+++ CHANGELOG.md
@@ -15,6 +15,14 @@ for the list of changes made since `v2.0.0-alpha.1`.
 
 ### Added
 
+-   New interval, month, and year helpers to fetch a list of all Saturdays and Sundays (weekends) for a given date interval. `eachWeekendOfInterval` is the handler function while the other two are wrapper functions.
+
+    -   `eachWeekendOfInterval`
+
+    -   `eachWeekendOfMonth`
+
+    -   `eachWeekendOfYear`
+
 - FP functions like those in [lodash](https://github.com/lodash/lodash/wiki/FP-Guide),
   that support [currying](https://en.wikipedia.org/wiki/Currying), and, as a consequence,
   functional-style [function composing](https://medium.com/making-internets/why-using-chain-is-a-mistake-9bc1f80d51ba).
@@ -698,7 +706,7 @@ for the list of changes made since `v2.0.0-alpha.1`.
 
 - `toDate` (previously `parse`) and `isValid` functions now accept `any` type
   as the first argument.
-  
+
 - [Exclude `docs.json` from the npm package](https://github.com/date-fns/date-fns/pull/837). Kudos to [@hawkrives](https://github.com/hawkrives).
 
 ### Fixed